create monopole_gr_base.hpp to avoid circular dependency, and delete some leftover tov code that doesn't belong
