+++ src/CMakeLists.txt
@@ -62,6 +62,7 @@ add_executable(phoebus
   microphysics/opac_phoebus/opac_phoebus.cpp
 
   monopole_gr/monopole_gr.hpp
+  monopole_gr/monopole_gr_base.hpp
   monopole_gr/monopole_gr_utils.hpp
   monopole_gr/monopole_gr.cpp
